140509_01/src/services/auth/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==1.10.15
email-validator==2.2.0
bcrypt==4.1.2
PyJWT==2.8.0
